Add ROPC OAuth2.0 grant
The only way to communicate with the v2 REST API is with requests authorized
by the OAuth2.0 protocol. As it stands, only the "Authorization Code" grant
and the "Resource Owner Password Credentials" grant are implemented in
MediaHaven.

Add functionality for the ROPC OAuth2.0 grant, leaving the door open for the
"Authorization Code" grant, providing an abstraction for requesting and
refreshing an auth token.