MHPY-13 return second response after token refresh
If a call is executed after the access token is expired, a new token is
requested with the refresh token. The same call is then executed but with
the new access token. However, the response of that second call didn't get
returned.

Bump responses to the latest version and use shorthand version to add
responses.