MHPY-15 Add tests
Add tests for `Records.create`.

Add type hints for frames and timecode arguments in the `Records.create`
function.

Change response of `MediaHavenClient._post` to actually return the content
of the body if there is any. Depending on the resource (path), the response
can contain content, for example a post request on the resource `exports`,
returns information of the created export.

Change `media_id` to `record_id` to reflect the signature and/or API path.