Support passing includeDeleted to the records endpoint
This boolean query parameter can be used to return items when they have
been logically deleted.  On the off-chance that capitalization matters
and to make it a little more ergonomic, we specifically handle the
parameter and convert it to a string value.  Additionally, we document
the other optional query parameter `fields'.