MHPY-11 Fix invalid refresh token
When the token is expired, a new one is requested given the refresh token.
If that renewal process fails because the refresh token is revoked or
expired, it should raise a RefreshTokenError by catching the underlying
oauthlib error.

In theory it should catch a InvalidGrantError (as defined in rfc6749):

"The provided authorization grant (e.g. authorization code, resource
owner credentials) or refresh token is invalid, expired, revoked, does
not match the redirection URI used in the authorization request, or was
issued to another client."

In practice an InvalidClientIdError is raised, which seems to be something
else: "Invalid client_id parameter value."

It is not entirely clear if this is because of oauthlib, oauthlib-requests,
the actual MediaHaven response or a combination.

As a solution, catch an InvalidClientIdError as well, and raise a
RefreshTokenError.